⚙️ add new theme parts for about page and widgets page that we insert instead of jer-intro and jer-widget-grid
It's convoluted, but has the effect that we can store these widgets in pages so the content persists across resets of the theme!

Ideally this should work a different way, e.g. with a custom post type where you create then choose the posts to show in a setting somewhere, but this one works well as long as you're ok hacking around.